task 5

+++ 0x0C-more_malloc_free/101-mul.c
@@ -29,7 +29,7 @@ int main(int argc, char *argv[])
 			}
 		}
 	}
-	multi =  atoi(argv[1]) * atoi(argv[2]);
+	multi =  atol(argv[1]) * atol(argv[2]);
 	printf("%lu\n", multi);
 
 	return (0);
